Add inputless datepicker
