+++ addon/templates/pikaday-inputless.hbs
@@ -1,0 +1,2 @@
+<input class="ember-pikaday-input" type="hidden">
+<div class="ember-pikaday-container"></div>
