+++ app/components/pikaday-inputless.js
@@ -1,0 +1,1 @@
+export { default } from 'ember-pikaday/components/pikaday-inputless';
