+++ tests/dummy/app/templates/application.hbs
@@ -16,3 +16,9 @@ Datepicker with bound and set value:
 
 Datepicker already disabled before it's rendered:
 {{pikaday-input disabled="disabled"}}
+
+<br>
+
+Inputless datepicker:
+{{pikaday-inputless onSelection=(action (mut someDate))}}
+{{someDate}}
